+++ core.impl/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/impl/jmx/DistributedJmxProxyFactoryImpl.java
@@ -845,6 +845,10 @@ public final class  DistributedJmxProxyFactoryImpl implements DistributedJmxProx
             {
                 throw new IllegalArgumentException("JMX Connection provider cannot be null");
             }
+            if (myIpTranslator == null)
+            {
+                throw new IllegalArgumentException("IpTranslator cannot be null");
+            }
             return new DistributedJmxProxyFactoryImpl(this);
         }
     }